+++ src/main/java/com/findmymeme/config/S3Config.java
@@ -9,9 +9,11 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 
 @Slf4j
 @Configuration
+@Profile("prod")
 public class S3Config {
 
     @Value("${cloud.aws.credentials.access-key}")
