+++ src/main/java/com/findmymeme/config/WebConfig.java
@@ -5,29 +5,10 @@ import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.support.ResourceBundleMessageSource;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
-
-    @Value("${file.upload.temp-dir}")
-    private String tempDir;
-
-    @Value("${file.upload.image-dir}")
-    private String imageDir;
-
-    @Value("${file.base-dir}")
-    private String baseDir;
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/temps/**")
-                .addResourceLocations("file:" + baseDir + "/" + tempDir + "/");
-
-        registry.addResourceHandler("/images/**")
-                .addResourceLocations("file:" + baseDir + "/" + imageDir + "/");
-    }
 
     @Bean
     public MessageSource messageSource(
