+++ src/main/java/com/findmymeme/file/service/S3FileStorageService.java
@@ -6,6 +6,7 @@ import com.findmymeme.exception.ErrorCode;
 import com.findmymeme.file.exception.FileStorageException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -13,6 +14,7 @@ import java.io.IOException;
 
 @Slf4j
 @Service
+@Profile("prod")
 public class S3FileStorageService implements FileStorageService {
     private static final String URL_FORMAT = "%s/%s";
     private static final String URL_SLASH = "/";
